requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
imapclient==3.0.1
email-validator==2.1.0